Update and add screen tests
